requirements.txt

fastapi==0.110.0
uvicorn==0.29.0
faiss-cpu==1.7.4
openai==1.16.1
sentence-transformers==2.6.1
python-dotenv==1.0.1
pydantic==2.6.4
httpx==0.27.0